Fix comments to be more style-compliant
Initials are not needed because git blame tells us who added the comment and associated code.

The new `retrieve_common_calendar_data` function was not documented.

Some comments that described the overall purpose of a function were not multiline and style compliant, so they could not be picked up by a linter.

Some comments were completely unnecessary.