Fixed the url issue
Fixed the url issue in the teams view to ensure that no one can edit the team, if they are not a member of it simply by just changing the url to edit. And I also did some changes with the 403, so now it uses the Djangos global handler403 approach just to make it reusable across the entire app.