reset button-settings page-minimum 3 characters popup
I changed the undo button to be a rest button that resets it too its original colours before any changes. i also centred the settings page and made the box it is larger to make it less crowded. and finally I did the task asking for the 3 minimum characters pop up to be removed